Move 'promises-util' to editor package to avoid dependency to core
