+++ packages/core/src/index.ts
@@ -3,4 +3,3 @@ export * from './ivy-script-client';
 export * from './inscription-client-jsonrpc';
 export * from './rcp-client';
 export * from './monaco-util';
-export * from './promises-util'
