+++ packages/editor/package.json
@@ -14,7 +14,6 @@
     "src"
   ],
   "dependencies": {
-    "@axonivy/inscription-core": "11.3.0-next",
     "@axonivy/inscription-protocol": "11.3.0-next",
     "@monaco-editor/react": "^4.6.0",
     "@radix-ui/react-accordion": "^1.1.2",
