+++ packages/editor/src/index.ts
@@ -5,3 +5,4 @@ export * from './context';
 export * from './monaco/monaco-editor-util';
 export * from './query';
 export * from './types/lambda';
+export * from './utils/promises-util'
